+++ CHANGELOG.md
@@ -1,8 +1,13 @@
 # edge-exchange-plugins
 
+# 0.5.4 (2019-04-09)
+
+- Add exchange rates from Coincap *legacy* API
+
 # 0.5.3 (2019-02-26)
 
 - Move ChangeNow into this repo for CORS reasons
+- Migrate Coincap to new API
 
 # 0.5.2 (2019-02-21)